refactor: 도메인 엔티티 및 리포지토리 수정

+++ storage/src/main/java/com/medicall/storage/db/core/hospital/HospitalRepository.java
@@ -1,8 +1,11 @@
 package com.medicall.storage.db.core.hospital;
 
 import java.util.List;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 
-public interface HospitalRepository extends JpaRepository<Hospital, Integer> {
+public interface HospitalRepository extends JpaRepository<Hospital, Long> {
     @Query("SELECT DISTINCT h FROM Hospital h " +
             "LEFT JOIN h.specialities s " +
             "LEFT JOIN s.major m " +
